External fonts and Logo added, changed the line height of the UL and navbar, resized and changed the color tones of the pictures for the Gallery, fixed the empty spaces between pics in gallery
Added two different google fonts. Increased the line height of the UL in order to add the logo and have more emphasis on the Logo by more negative space around it. Resized and fixed the color property of the pictures in order to be more appealing and interesting. Added clearfix to the grid to omit the dead space between the pictures.